# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_0xsj_nexus-go_internal_flags_application_dto.CreateFlagRequest:
    properties:
      description:
        maxLength: 1024
        type: string
      enabled:
        type: boolean
      key:
        maxLength: 128
        minLength: 1
        type: string
      name:
        maxLength: 255
        type: string
    required:
    - key
    type: object
  github_com_0xsj_nexus-go_internal_flags_application_dto.CreateFlagResponse:
    properties:
      flag:
        $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.FlagDTO'
    type: object
  github_com_0xsj_nexus-go_internal_flags_application_dto.DeleteFlagResponse:
    properties:
      id:
        type: string
      success:
        type: boolean
    type: object
...
  github_com_0xsj_nexus-go_internal_flags_application_dto.ListFlagsResponse:
    properties:
      flags:
        items:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.FlagSummaryDTO'
        type: array
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  internal_flags_interface_http_v1.ErrorResponse:
    description: Error response returned by the API
    properties:
      code:
        example: FLAG_NOT_FOUND
        type: string
      message:
        example: flag not found
        type: string
      meta:
        additionalProperties: {}
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Multi-tenant API built with hexagonal architecture. Provides identity
    management, tenant management, email templates, and more.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Hexagonal Go API
  version: "1.0"
paths:
...
  /api/v1/flags:
    get:
      description: Retrieves a paginated list of feature flags
      parameters:
      - description: Filter by enabled status
        in: query
        name: enabled
        type: boolean
      - description: Search by key, name, or description
        in: query
        name: search
        type: string
      - default: 50
        description: Pagination limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of flags
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.ListFlagsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List feature flags
      tags:
      - flags
    post:
      consumes:
      - application/json
      description: Creates a new feature flag
      parameters:
      - description: Create flag request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.CreateFlagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Flag created successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.CreateFlagResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "409":
          description: Flag with key already exists
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create feature flag
      tags:
      - flags
  /api/v1/flags/{id}:
    delete:
      description: Deletes a feature flag
      parameters:
      - description: Flag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Flag deleted successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.DeleteFlagResponse'
        "400":
          description: Invalid flag ID
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "404":
          description: Flag not found
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete feature flag
      tags:
      - flags
    get:
      description: Retrieves a feature flag by its unique identifier
      parameters:
      - description: Flag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Flag found
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.GetFlagResponse'
        "400":
          description: Invalid flag ID
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "404":
          description: Flag not found
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get feature flag by ID
      tags:
      - flags
    put:
      consumes:
      - application/json
      description: Updates an existing feature flag
      parameters:
      - description: Flag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Update flag request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.UpdateFlagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Flag updated successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.UpdateFlagResponse'
        "400":
          description: Invalid request body or flag ID
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "404":
          description: Flag not found
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update feature flag
      tags:
      - flags
  /api/v1/flags/{id}/disable:
    post:
      description: Disables a feature flag
      parameters:
      - description: Flag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Flag disabled successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.DisableFlagResponse'
        "400":
          description: Invalid flag ID
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "404":
          description: Flag not found
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disable feature flag
      tags:
      - flags
  /api/v1/flags/{id}/enable:
    post:
      description: Enables a feature flag
      parameters:
      - description: Flag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Flag enabled successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.EnableFlagResponse'
        "400":
          description: Invalid flag ID
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "404":
          description: Flag not found
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enable feature flag
      tags:
      - flags
  /api/v1/flags/{id}/overrides:
    delete:
      consumes:
      - application/json
      description: Removes an override for a specific target
      parameters:
      - description: Flag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Remove override request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.RemoveOverrideRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Override removed successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.RemoveOverrideResponse'
        "400":
          description: Invalid request body or flag ID
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "404":
          description: Flag or override not found
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove override
      tags:
      - flags
    post:
      consumes:
      - application/json
      description: Sets an override for a specific target (tenant or user)
      parameters:
      - description: Flag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Set override request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.SetOverrideRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Override set successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.SetOverrideResponse'
        "400":
          description: Invalid request body or flag ID
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "404":
          description: Flag not found
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set override
      tags:
      - flags
  /api/v1/flags/{id}/rules:
    post:
      consumes:
      - application/json
      description: Adds a targeting rule to a feature flag
      parameters:
      - description: Flag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Add rule request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.AddRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Rule added successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.AddRuleResponse'
        "400":
          description: Invalid request body or flag ID
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "404":
          description: Flag not found
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add targeting rule
      tags:
      - flags
  /api/v1/flags/{id}/rules/{ruleId}:
    delete:
      description: Removes a targeting rule from a feature flag
      parameters:
      - description: Flag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Rule ID
        format: uuid
        in: path
        name: ruleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Rule removed successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.RemoveRuleResponse'
        "400":
          description: Invalid flag ID or rule ID
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "404":
          description: Flag or rule not found
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove targeting rule
      tags:
      - flags
...
  /api/v1/flags/by-key:
    get:
      description: Retrieves a feature flag by its key
      parameters:
      - description: Flag key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Flag found
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_flags_application_dto.GetFlagResponse'
        "400":
          description: Key is required
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
        "404":
          description: Flag not found
          schema:
            $ref: '#/definitions/internal_flags_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get feature flag by key
      tags:
      - flags
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...